csv parser

+++ app/controllers/houses_controller.rb
@@ -165,9 +165,17 @@ class HousesController < ApplicationController
 
     csvparse = CsvParser.new((csvfile),params[:county])
     csvparse.run
-    session[:house_attributes] = csvparse.gethouse.attributes
-    session[:sales] = csvparse.getsales
-    File.delete(csvfile)
-    redirect_to new_house_path
+    if csvparse.house_exists
+      session[:sales] = csvparse.get_sales
+      house = csvparse.get_house
+      flash[:notice] = "House Already Exists. Sales Information Does Not."
+      redirect_to new_house_sale_path(house)
+    else
+      session[:house_attributes] = csvparse.get_house.attributes
+      session[:sales] = csvparse.get_sales
+      File.delete(csvfile)
+      redirect_to new_house_path
+    end
+
   end
 end
